requirements.txt
numpy==2.0.2
openai-whisper==20240930
pynput==1.8.1
sounddevice==0.5.1
torch==2.6.0